Move use of funds jsx and styles to shared
