+++ desktop/styles/styles_ctrl.pcss
@@ -10,6 +10,6 @@
 @import 'BannerText.pcss';
 @import './ProgressBar.pcss';
 @import './BackLink.pcss';
-@import './FundsModal.pcss';
-@import './FundsDistributionInfo.pcss';
-@import './CompanyBudgets.pcss';
+@import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
+@import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
+@import '../../shared/components/ui/use_of_funds/CompanyBudgets.pcss';
