+++ shared/components/ui/use_of_funds/FundsDistributionInfo.pcss
